requirements.txt
openai
python-dotenv
playwright
asyncpg
